munin: exclude code cache from heap calculation (java>=8)
Code cache (`memory['code']`) is not part of Java heap, as is
reported by runtime (`'is_heap': False` for all code related
records).

I've also seen the value of free Java heap turn negative for
roughly the value of code cache when eden space usage is at its
peak, which also screams that code cache should be excluded.

I don't know if it's also applicable for java_version < 8, but
hopefully it'll go away soon(ish) and we can leave it as is.